perf(canvas-workspace): React.lazy the 5 heavy node bodies, evict tiptap extensions + lowlight from entry (C1/C6)
DefaultCanvasNode now React.lazy-loads the 5 heavy node bodies
(text/file/agent/frame/terminal) behind a single <Suspense fallback={null}>
inside the .node-body div; the 3 light bodies (iframe/dynamic-app/plugin)
stay static. To make the heavy bodies cleanly lazy, two sub-exports were
extracted out of the always-on CanvasNodeHeader's reach:
  - TextColorPicker -> TextNodeBody/TextColorPicker.tsx (tiptap-free)
  - FrameChildrenToggle + FrameColorPicker -> FrameNodeBody/FrameHeaderControls.tsx
    (xterm-free; breaks the header -> FrameNodeBody -> AgentTeamFrame ->
    AgentNodeBody -> xterm always-on chain)

Entry chunk: 4223 -> 3219 KB raw (-23.8%) / 928 -> 706 KB gzip (-23.9%).
Evicted into lazy chunks: @tiptap/starter-kit + the extension pack + lowlight
(35 hljs grammars) + tiptap-markdown, loaded on first mount of a text/file
node. perf:bundle 3/3 gates pass post-ratchet.

NOT evicted this PR (follow-ups, all in entry by design):
  - @tiptap/react + @tiptap/pm via chain B (Canvas -> useCanvasSearch ->
    noteSearchExtension; noteSearchPluginKey is a module-load PluginKey,
    unbreakable without a FileNodeEditorRegistry refactor)
  - @xterm/xterm via WorkspaceTerminalDock (C2)
  - highlight.js via chat/utils/markdown.ts (C3)
  - module-federation runtime (C7)

Guards (ratchet rule — same-PR baseline + WATCHLIST):
  - bundle-boundaries.test.ts: WATCHLIST += @tiptap/starter-kit, lowlight
    (verified out of the static graph). @tiptap/react + @tiptap/pm deliberately
    NOT watchlisted — still static via chain B. EXPECTED_STATIC unchanged.
  - collect-metrics.mjs: bundle.lazy_boundary_watchlist generalized to
    mermaid + force-graph (force-graph was lazy since B6 but ungated here —
    consistency fix). lowlight gated via the static-graph test, not a probe
    (its 'expected `grammar`' string is tree-shaken, unusable).
  - baselines.json: entryRawKB 4223->3219, entryGzipKB 928->706 (totalJsKB
    unchanged — code relocates, not removed).

Verified: 7/7 bundle-boundaries tests, typecheck clean, 665/666 canvas-workspace
tests (the 1 failure is pre-existing file-size-governance baseline drift,
unrelated — confirmed by stash on clean HEAD). Adversarial verify workflow
independently confirmed eviction claims (entry probe-grep + static-graph walk)
and Suspense safety (no forwardRef/useImperativeHandle; props-only dispatch;
canvas keepAlive; .node-body flex:1 preserves height during fallback).

Known tradeoff: TextNodeBody auto-focus on the first empty text node of a
session is deferred behind a ~114KB-gzip chunk load (one-time; keepAlive
caches it). Recommended non-blocking follow-up: idle-prefetch the text body
chunk on canvas mount.
